Changes for new version 4.0.0
