+++ src/framework/arduino/examples/E9000SPC/E9000SPC.ino
@@ -1,8 +1,8 @@
 /*!
  * \name        E9000SPC
  * \author      Infineon Technologies AG
- * \copyright   2020 Infineon Technologies AG
- * \version     3.1.0
+ * \copyright   2020-2024 Infineon Technologies AG
+ * \version     4.0.0
  * \brief       This example shows how to handle the TLE5012B-E9000 variant with SPC interface
  * \details
  * The TLE5012B-E9000 with SPC interface does not start the DSP automatically in a loop at start up.
@@ -23,7 +23,7 @@
  *
  */
 
-#include <TLE5012-ino.hpp>
+#include <tlx5012-arduino.hpp>
 
 using namespace tle5012;
 
@@ -46,10 +46,13 @@ void setup() {
   while (!Serial) {};
 
   checkError = Tle5012Sensor.begin();
-  Serial.println("init done!");
+  Serial.print("checkError: ");
+  Serial.println(checkError, HEX);
+  delay(1000);
 
   Tle5012Sensor.resetFirmware();
   Tle5012Sensor.sBus->triggerUpdate();
+  Serial.println("init done!");
   delay(1000);
 
   pinMode(IFA, OUTPUT);
