+++ src/framework/arduino/examples/readAngleValueProcessing/readAngleValueProcessing.ino
@@ -1,8 +1,8 @@
 /*!
  * \name        readAngleValueProcessing
  * \author      Infineon Technologies AG
- * \copyright   2020 Infineon Technologies AG
- * \version     3.1.0
+ * \copyright   2020-2024 Infineon Technologies AG
+ * \version     4.0.0
  * \brief       Reads angle values
  * \details
  * This is a fast running example to demonstrate the possible readout
@@ -19,7 +19,7 @@
  *
  */
 
-#include <TLE5012-ino.hpp>
+#include <tlx5012-arduino.hpp>
 
 using namespace tle5012;
 
@@ -32,7 +32,7 @@ void setup() {
   while (!Serial) {};
   Serial.println("init done!");
   checkError = Tle5012Sensor.begin();
-  Serial.print("checkerror: ");
+  Serial.print("checkError: ");
   Serial.println(checkError, HEX);
   delay(1000);
 }
