+++ src/framework/arduino/examples/sensorType/const.h
@@ -1,8 +1,8 @@
 /*!
  * \name        const
  * \author      Infineon Technologies AG (Dr.Olaf Filies)
- * \copyright   2020 Infineon Technologies AG
- * \version     3.1.0
+ * \copyright   2020-2024 Infineon Technologies AG
+ * \version     4.0.0
  * \brief       macros and other const variables
  *
  * \attention
@@ -62,7 +62,7 @@
 #define sc_SPCpin       F("\nIFAB IFA pin:           \t")
 
 // additional
-#define sc_MOD2cal        F("\nMOD2 Autocalibration:   \t")
+#define sc_MOD2cal        F("\nMOD2 auto calibration:  \t")
 #define sc_MOD2predict    F("\nMOD2 Prediction:        \t")
 #define sc_MODpin         F("\nMOD3 IFA/B/C pin set:   \t")
 #define sc_MODstrongfast  F("strong driver, DATA: strong driver, fast edge")
