+++ src/framework/arduino/pal/pal-gpio-arduino.hpp
@@ -1,8 +1,9 @@
 /**
- * @file        gpio-arduino.hpp
- * @brief       Arduino PAL for the GPIO
- * @date        May 2020
- * @copyright   Copyright (c) 2019-2020 Infineon Technologies AG
+ * \file        pal-gpio-arduino.hpp
+ * \brief       Arduino PAL for the GPIO
+ * \author      Infineon Technologies AG
+ * \version     4.0.0
+ * \copyright   2020-2024 Infineon Technologies AG
  *
  * SPDX-License-Identifier: MIT
  */
@@ -11,12 +12,8 @@
 #ifndef PAL_GPIO_ARDUINO_HPP_
 #define PAL_GPIO_ARDUINO_HPP_
 
-#include "../../../config/tle5012-conf.hpp"
 
-#if (TLE5012_FRAMEWORK == TLE5012_FRMWK_ARDUINO)
-
-//#include "../../../corelib/TLE5012b.hpp"
-#include "../../../pal/pal-gpio.hpp"
+#include "tlx5012-pal-gpio.hpp"
 
 
 namespace tle5012
@@ -33,29 +30,28 @@ namespace tle5012
 	 */
 	class GPIOIno: virtual public GPIOPAL
 	{
-	private:
-		uint8_t     pin;
-		uint8_t     mode;
-		VLogic_t    logic;
+		private:
+			uint8_t     pin;
+			uint8_t     mode;
+			VLogic_t    logic;
 
-	public:
-		static constexpr uint8_t  UNUSED_PIN = 0xFFU;    /**< Unused pin */
+		public:
+			static constexpr uint8_t  UNUSED_PIN = 0xFFU;    /**< Unused pin */
 
-					GPIOIno();
-					GPIOIno(uint8_t pin, uint8_t mode, VLogic_t logic);
-					~GPIOIno();
-		Error_t     init();
-		Error_t     changeMode(uint8_t mode);
-		Error_t     deinit();
-		VLevel_t    read();
-		Error_t     write(VLevel_t level);
-		Error_t     enable();
-		Error_t     disable();
+						GPIOIno();
+						GPIOIno(uint8_t pin, uint8_t mode, VLogic_t logic);
+						~GPIOIno();
+			Error_t     init();
+			Error_t     changeMode(uint8_t mode);
+			Error_t     deinit();
+			VLevel_t    read();
+			Error_t     write(VLevel_t level);
+			Error_t     enable();
+			Error_t     disable();
 	};
 
 	/** @} */
 
 }
 
-#endif /** TLE5012_FRAMEWORK **/
 #endif /** PAL_GPIO_ARDUINO_HPP_ **/
