+++ src/framework/arduino/pal/spi3w-esp32.cpp
@@ -1,17 +1,16 @@
 /*
- * @file        spi-platf-ino.cpp
- * @brief       TLE5012 Arduino SPI class wrapper
- * @date        October 2020
- * @copyright   Copyright (c) 2019-2020 Infineon Technologies AG
- *
+ * \file        spi3w-esp32-ino.cpp
+ * \brief       ESP32 3wire SPI cover
+ * \author      Infineon Technologies AG
+ * \version     4.0.0
+ * \copyright   2020-2024 Infineon Technologies AG
+ * 
  * SPDX-License-Identifier: MIT
  */
 
 #include "spi3w-ino.hpp"
 #include "pal-pin-types.hpp"
 
-//
-#if (TLE5012_FRAMEWORK == TLE5012_FRMWK_ARDUINO)
 #if (SPI3W_INO == SPI3W_ESP32)
 
 #include <driver/spi_master.h>
@@ -181,4 +180,3 @@ void SPIClass3W::sendReceiveSpi(uint16_t *sent_data, uint16_t size_of_sent_data,
 /** @} */
 
 #endif /* SPI3W_INO */
-#endif /* TLE5012_FRAMEWORK */
